requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-cors
flask-jwt-extended
python-dotenv
mysqlclient
pyjwt
werkzeug